requirements.txt
annotated-types==0.5.0
anyio==3.7.1
banditsdk==0.2.1
click==8.1.6
exceptiongroup==1.1.2
fastapi==0.100.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
loguru==0.7.0
Pillow==10.0.0
pydantic==2.1.1
pydantic_core==2.4.0
python-multipart==0.0.6
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.1
zipp==3.16.2
